+++ test/fixtures/User.js
@@ -1,27 +1,36 @@
 module.exports = [
     {
         id: 'c', // auto-generated, but still string required
+        role: 'admin',
         firstName: 'NeoNexus',
         lastName: 'DeMortis',
-        role: 'admin',
         email: 'admin@testing.com',
         password: 'Some really long passphrase is better than a password.',
         verificationKey: 'rmWFXL/4wePcEqrojD6IGgzGnlQfGlaZUIoe7mWYbTQ='
     }, {
         id: 'c', // auto-generated, but still string required
+        role: 'manager',
         firstName: 'Manager',
         lastName: 'Bot',
-        role: 'manager',
         email: 'manager@testing.com',
         password: `Passphrases are great, aren't they?`,
         verificationKey: 'iQXYf9NeNbzlB6qeIOqtiGNRyPdM7y/AyDf1rHYRi78='
     }, {
         id: 'c', // auto-generated, but still string required
+        role: 'user',
         firstName: 'User',
         lastName: 'Bot',
-        role: 'user',
         email: 'user@testing.com',
         password: `Who doesn't love a good passphrase?`,
         verificationKey: 'O+mATUT1soiuaqAfS9XJcWkQscs/xI/IaYq37KLCIrw='
+    }, {
+        id: 'c', // auto-generated, but still string required
+        role: 'user',
+        firstName: 'Deleted',
+        lastName: 'Softly',
+        email: 'softly@deleted.com',
+        password: `Mmmmmmmmmmmmmmm, tasty passphrase...`,
+        verificationKey: 'zxJQsKi85AcpN+CmSisE9fJuDApd61mgNKptUj5sfsE=',
+        deletedAt: new Date()
     }
 ];